backend/pyproject.toml
[project]
name = "smartb-backend"
version = "0.1.0"
description = "FastAPI backend with JWT auth, RBAC context hierarchy, and PostgreSQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.35",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.14.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib>=1.7.4",
  "pydantic-settings>=2.6.1",
  "python-multipart>=0.0.9",
  "email-validator>=2.2.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
